I've moved the code for the Gmail API and the scraping code to their own individual files. I have managed to get the Gmail API to work and I am now able to send emails across. I need to include attachments to the email such as an image.
