+++ server/token.json
@@ -1,1 +1,1 @@
-{"access_token":"ya29.GltWBxTVLG1rHRKhtt041pETXcZNKXYfuezcM8kLiSquov87ha_fYfp5LuiqVJMFatO962vm7fCVr6fuOzMgpnw9bGLlP010CtI4PwAKi_ugUQgdEvyuzKem1xat","refresh_token":"1/ZOFcCsRUNevz6Vb0hR8_nLn4VCSxgMQ8a63EEd-PPdY","scope":"https://www.googleapis.com/auth/gmail.readonly","token_type":"Bearer","expiry_date":1564604943504}
+{"access_token":"ya29.GltYB63kM9k4b8lcUBIJja9QlDsng6wTjRcvf50sR94uRI9MaPF7N-6IMtzZthQoPpoe3LeU9sXOBsC0iT07_QCIZQbtm31nza-j9zHtWKyeXWqmnzM-21XfenM8","refresh_token":"1/g_pOehgahx22skK53oMGRqZA0K-hv3111n42RK-1A1U","scope":"https://mail.google.com/ https://www.googleapis.com/auth/gmail.modify https://www.googleapis.com/auth/gmail.send https://www.googleapis.com/auth/gmail.compose","token_type":"Bearer","expiry_date":1564772775440}
